+++ lib/applePay.ts
@@ -30,16 +30,37 @@ const DEFAULT_CONFIG = {
   },
 }
 
-// Starts the Apple Pay session, registers event handlers
-function startApplePaySession(config: any, apiKey: string): void {
+let tokens: ApplePayJS.ApplePayPaymentToken
+
+interface PaymentToken {
+  version: string
+  data: string
+  signature: string
+  header: {
+    ephemeralPublicKey: string
+    publicKeyHash: string
+    transactionId: string
+  }
+}
+
+// Starts the Apple Pay session to pay on backend, registers event handlers
+function startApplePaySessionBackendPay(config: any, apiKey: string): void {
   const applePaySession: ApplePaySession = new ApplePaySession(6, config)
-  handleApplePayEvents(applePaySession, apiKey)
+  handleCommonApplePayEvents(applePaySession)
+  handleApplePayPaymentOnBackendEvent(applePaySession, apiKey)
+  applePaySession.begin()
+}
+
+function startApplePaySessionFrontendPay(config: any): void {
+  const applePaySession: ApplePaySession = new ApplePaySession(6, config)
+  handleCommonApplePayEvents(applePaySession)
+  handleApplePayPaymentOnFrontendEvent(applePaySession)
   applePaySession.begin()
 }
 
 // Registers event handlers. There are 5 steps associated with Apple Pay, transition between steps triggers event, these events are:
 // onvalidatemerchant, onshippingcontactselected, onshippingmethodselected, completeShippingMethodSelection and onpaymentauthorized.
-function handleApplePayEvents(appleSession: ApplePaySession, apiKey: string) {
+function handleCommonApplePayEvents(appleSession: ApplePaySession) {
   // This is the first event that Apple triggers. Validate the Apple Pay Session from endpoint.
   appleSession.onvalidatemerchant = function (
     event: ApplePayJS.ApplePayValidateMerchantEvent
@@ -124,9 +145,14 @@ function handleApplePayEvents(appleSession: ApplePaySession, apiKey: string) {
       newLineItems // order summary updated
     )
   }
+}
 
-  // Final event which performs debit. Triggered after the user has confirmed the transaction with the Touch ID or Face ID.
-  // All details about the customer are provided and most importantly we get the Apple Pay token in payload needed to perform a payment.
+// Final event which performs debit. Triggered after the user has confirmed the transaction with the Touch ID or Face ID.
+// All details about the customer are provided and most importantly we get the Apple Pay token in payload needed to perform a payment.
+function handleApplePayPaymentOnBackendEvent(
+  appleSession: ApplePaySession,
+  apiKey: string
+) {
   appleSession.onpaymentauthorized = function (
     event: ApplePayJS.ApplePayPaymentAuthorizedEvent
   ) {
@@ -143,6 +169,17 @@ function handleApplePayEvents(appleSession: ApplePaySession, apiKey: string) {
         appleSession.completePayment(ApplePaySession.STATUS_FAILURE)
       }
     })
+  }
+}
+
+function handleApplePayPaymentOnFrontendEvent(appleSession: ApplePaySession) {
+  appleSession.onpaymentauthorized = function (
+    event: ApplePayJS.ApplePayPaymentAuthorizedEvent
+  ) {
+    console.log('received authorization')
+    console.log(event.payment)
+    console.log(JSON.stringify(event.payment.token))
+    tokens = event.payment.token
   }
 }
 
@@ -202,4 +239,15 @@ function applePayAvailable(): boolean {
   // canMakePaymentsWithActiveCard(merchantIdentifier: string):
 }
 
-export { startApplePaySession, DEFAULT_CONFIG, applePayAvailable }
+function getApplePayTokens() {
+  return tokens
+}
+
+export {
+  startApplePaySessionBackendPay,
+  startApplePaySessionFrontendPay,
+  DEFAULT_CONFIG,
+  applePayAvailable,
+  getApplePayTokens,
+  PaymentToken,
+}
